requirements.txt
python-binance==1.0.19
numpy==2.0.2
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
openai==1.35.13
httpx==0.24.1
pymongo==4.6.0
pyyaml==6.0.3
boto3==1.40.71
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
pytz>=2023.3
beautifulsoup4>=4.12.0